Finally got it to work! I had to download a WarZombie!
NPCMovement.cs added an unused WaypointManager and a Animator. Populated them in the Start method, used State management via the update method, Follow, SetIdle, Attack and Search are the States. Also added a GetRandomPositionOnMap which ensures that random position is indeed on the map.
RaycastShoot.cs added the field AimInaccuracy to add a little inaccurate shooting. Added WaterCircleRadius to stop spawning enemies within the water feature.
Enemy.cs added a call to disable the Animator in the TriggerRagdoll.